+++ packages/react-three-rapier/src/shared-objects.ts
@@ -5,3 +5,6 @@ export const _euler = new Euler();
 export const _vector3 = new Vector3();
 export const _object3d = new Object3D();
 export const _matrix4 = new Matrix4();
+export const _position = new Vector3();
+export const _rotation = new Quaternion();
+export const _scale = new Vector3();
